LayoutTests/http/tests/websocket/tests/hybi/close-on-unload-reference-in-parent-expected.txt

Test if Web Socket opened in iframe but referred in parent document is closed on unloading the frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws on master document is ready.
PASS insert a iframe, where open ws called 'socket1'
PASS document.childWebSocket is non-null.
PASS 'socket1' is sent to the server. unload the iframe. expect receiving 'socket1' on ws...
PASS closedSocket is "socket1: receive next message"
PASS document.childWebSocket.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
